Move renderAction to base hover widget
